Refactor: move GSP endpoint to routes/gsp/, extract graph resolution, fix repo metadata query
- Move collection /graph GSP endpoint logic from Collections.kt to routes/gsp/CollectionGraphRead.kt
- Extract graph resolution SPARQL into reusable resolveCollectionGraphIris() in CollectionGraphResolution.kt
- Fix graph resolution query to properly resolve refs through repo metadata graphs:
  - Branch: query repo metadata graph for commit -> staging snapshot -> graph
  - Lock: query repo metadata graph for commit -> model snapshot -> graph
  - Scratch: derive graph IRI from repo path + scratch ID
- Both CollectionQuery.kt and CollectionGraphRead.kt now use the shared utility
- Collections.kt delegates to readCollectionGraph() following the same pattern as Model.kt
